# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Heiba private video streaming platform API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token in the Authorization header
    cookieAuth:
      type: apiKey
      in: cookie
      name: heiba_session
      description: Session token in an httpOnly cookie
  schemas:
...
    UserRole:
      type: string
      enum: [OWNER, ADMIN, GROUP_MANAGER, MEMBER]
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    GroupRole:
      type: string
      enum: [manager, member]
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        ownerId:
          type: integer
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - ownerId
        - memberCount
        - createdAt
        - updatedAt
    GroupListResult:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
        total:
          type: integer
      required:
        - groups
        - total
    CreateGroupInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
      required:
        - name
    UpdateGroupInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
          nullable: true
    GroupMember:
      type: object
      properties:
        userId:
          type: integer
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        roleInGroup:
          $ref: "#/components/schemas/GroupRole"
        joinedAt:
          type: string
          format: date-time
      required:
        - userId
        - email
        - role
        - roleInGroup
        - joinedAt
    GroupMemberListResult:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupMember"
      required:
        - members
    AddGroupMemberInput:
      type: object
      properties:
        userId:
          type: integer
        roleInGroup:
          $ref: "#/components/schemas/GroupRole"
      required:
        - userId
    UpdateGroupMemberInput:
      type: object
      properties:
        roleInGroup:
          $ref: "#/components/schemas/GroupRole"
      required:
        - roleInGroup
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List groups visible to the current user
      description: OWNER/ADMIN see all groups; GROUP_MANAGER sees groups they manage; MEMBER sees groups they belong to.
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Groups list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupListResult"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group (OWNER/ADMIN)
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupInput"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Group name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a group (members, managers, OWNER/ADMIN)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Group not found or not visible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update a group (OWNER/ADMIN)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupInput"
      responses:
        "200":
          description: Group updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Group name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group (OWNER/ADMIN)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Group deleted
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /groups/{id}/members:
    get:
      operationId: listGroupMembers
      tags: [groups]
      summary: List group members (members, managers, OWNER/ADMIN)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMemberListResult"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Group not found or not visible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: addGroupMember
      tags: [groups]
      summary: Add a member to a group (OWNER/ADMIN, or manager of the group)
      description: Managers can only add members with the "member" group role; assigning "manager" requires OWNER/ADMIN.
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddGroupMemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMember"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Group or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User is already a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /groups/{id}/members/{userId}:
    patch:
      operationId: updateGroupMember
      tags: [groups]
      summary: Change a member's role within the group (OWNER/ADMIN for "manager"; managers can only toggle members)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupMemberInput"
      responses:
        "200":
          description: Member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMember"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Membership not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: removeGroupMember
      tags: [groups]
      summary: Remove a member from a group (OWNER/ADMIN, or manager of the group)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Member removed
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Membership not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...