# lib/api-zod/src/generated/types/addGroupMemberInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Heiba private video streaming platform API
 * OpenAPI spec version: 0.1.0
 */
import type { GroupRole } from './groupRole';

export interface AddGroupMemberInput {
  userId: number;
  roleInGroup?: GroupRole;
}