# lib/api-zod/src/generated/types/updateGroupInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Heiba private video streaming platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateGroupInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name?: string;
  /** @maxLength 2000 */
  description?: string | null;
}